requirements.txt
Django==1.10.3
django-cors-headers==1.2.2
django-filter==0.15.3
djangorestframework==3.5.2
Markdown==2.6.7
